Move init_sct to spinalcordtoolbox.utils (#2943)
* Move init_sct to sct.utils

* Refactor callers of init_sct()

* init_sct: add forgotten add_elapsed_time_counter() and init_error_client()

* code review fixups
