+++ scripts/sct_deepseg.py
@@ -18,18 +18,16 @@ import argparse
 import os
 import sys
 
-import spinalcordtoolbox as sct
-import spinalcordtoolbox.deepseg.core
-import spinalcordtoolbox.deepseg.models
-import spinalcordtoolbox.utils
-from sct_utils import init_sct, printv, display_viewer_syntax
+import spinalcordtoolbox.deepseg as deepseg
+from spinalcordtoolbox.utils import SmartFormatter, Metavar, init_sct
+from sct_utils import printv, display_viewer_syntax
 
 
 def get_parser():
     parser = argparse.ArgumentParser(
         description="Segment an anatomical structure or pathologies according to the specified deep learning model.",
         add_help=None,
-        formatter_class=sct.utils.SmartFormatter,
+        formatter_class=SmartFormatter,
         prog=os.path.basename(__file__).strip(".py"))
 
     input_output = parser.add_argument_group("\nINPUT/OUTPUT")
@@ -37,19 +35,19 @@ def get_parser():
         "-i",
         required=True,
         help="Image to segment.",
-        metavar=sct.utils.Metavar.file)
+        metavar=Metavar.file)
     input_output.add_argument(
         "-o",
         help="Output file name. In case of multi-class segmentation, class-specific suffixes will be added. By default,"
              "suffix '_seg' will be added and output extension will be .nii.gz.",
-        metavar=sct.utils.Metavar.str)
+        metavar=Metavar.str)
 
     seg = parser.add_argument_group('\nTASKS')
     seg.add_argument(
         "-task",
         help="Task to perform. It could either be a pre-installed task, task that could be installed, or a custom task."
              " To list available tasks, run: sct_deepseg -list-tasks",
-        metavar=sct.utils.Metavar.str)
+        metavar=Metavar.str)
     seg.add_argument(
         "-list-tasks",
         action='store_true',
@@ -57,7 +55,7 @@ def get_parser():
     seg.add_argument(
         "-install-task",
         help="Install models that are required for specified task.",
-        choices=list(sct.deepseg.models.TASKS.keys()))
+        choices=list(deepseg.models.TASKS.keys()))
 
     seg = parser.add_argument_group('\nMODELS')
     seg.add_argument(
@@ -68,7 +66,7 @@ def get_parser():
              "-model my_models/model. To list official models, run: sct_deepseg -list-models."
              "To build your own model, follow instructions at: https://github.com/neuropoly/ivado-medical-imaging",
         nargs='+',
-        metavar=sct.utils.Metavar.str)
+        metavar=Metavar.str)
     seg.add_argument(
         "-list-models",
         action='store_true',
@@ -76,7 +74,7 @@ def get_parser():
     seg.add_argument(
         "-install-model",
         help="Install specified model.",
-        choices=list(sct.deepseg.models.MODELS.keys()))
+        choices=list(deepseg.models.MODELS.keys()))
 
     misc = parser.add_argument_group('\nPARAMETERS')
     misc.add_argument(
@@ -85,20 +83,20 @@ def get_parser():
         help="Binarize segmentation with specified threshold. Set to 0 for no thresholding (i.e., soft segmentation). "
              "Default value is model-specific and was set during optimization "
              "(more info at https://github.com/sct-pipeline/deepseg-threshold).",
-        metavar=sct.utils.Metavar.float,
-        default=sct.deepseg.core.DEFAULTS['thr'])
+        metavar=Metavar.float,
+        default=deepseg.core.DEFAULTS['thr'])
     misc.add_argument(
         "-largest",
         type=int,
         help="Keep the largest connected-objects from the output segmentation. Specify the number of objects to keep."
              "To keep all objects, set to 0",
-        default=sct.deepseg.core.DEFAULTS['largest'])
+        default=deepseg.core.DEFAULTS['largest'])
     misc.add_argument(
         "-fill-holes",
         type=int,
         help="Fill small holes in the segmentation.",
         choices=(0, 1),
-        default=sct.deepseg.core.DEFAULTS['fill_holes'])
+        default=deepseg.core.DEFAULTS['fill_holes'])
 
     misc = parser.add_argument_group('\nMISC')
     misc.add_argument(
@@ -122,19 +120,19 @@ def main():
 
     # Deal with model
     if args.list_models is not None:
-        sct.deepseg.models.display_list_models()
+        deepseg.models.display_list_models()
 
     # Deal with task
     if args.list_tasks is not None:
-        sct.deepseg.models.display_list_tasks()
+        deepseg.models.display_list_tasks()
 
     if args.install_model is not None:
-        sct.deepseg.models.install_model(args.install_model)
+        deepseg.models.install_model(args.install_model)
         exit(0)
 
     if args.install_task is not None:
-        for name_model in sct.deepseg.models.TASKS[args.install_task]['models']:
-            sct.deepseg.models.install_model(name_model)
+        for name_model in deepseg.models.TASKS[args.install_task]['models']:
+            deepseg.models.install_model(name_model)
         exit(0)
 
     # Deal with input/output
@@ -147,7 +145,7 @@ def main():
 
     # Get pipeline model names
     if args.task is not None:
-        name_models = sct.deepseg.models.TASKS[args.task]['models']
+        name_models = deepseg.models.TASKS[args.task]['models']
 
     if args.model is not None:
         name_models = args.model
@@ -156,7 +154,7 @@ def main():
     fname_prior = None
     for name_model in name_models:
         # Check if this is an official model
-        if name_model in list(sct.deepseg.models.MODELS.keys()):
+        if name_model in list(deepseg.models.MODELS.keys()):
             # If it is, check if it is installed
             path_model = spinalcordtoolbox.deepseg.models.folder(name_model)
             if not spinalcordtoolbox.deepseg.models.is_valid(path_model):
@@ -165,11 +163,11 @@ def main():
         # If it is not, check if this is a path to a valid model
         else:
             path_model = os.path.abspath(name_model)
-            if not sct.deepseg.models.is_valid(path_model):
+            if not deepseg.models.is_valid(path_model):
                 parser.error("The input model is invalid: {}".format(path_model))
 
         # Call segment_nifti
-        fname_seg = sct.deepseg.core.segment_nifti(args.i, path_model, fname_prior, args)
+        fname_seg = deepseg.core.segment_nifti(args.i, path_model, fname_prior, args)
         # Use the result of the current model as additional input of the next model
         fname_prior = fname_seg
