+++ scripts/sct_dmri_compute_bvalue.py
@@ -21,7 +21,7 @@ import math
 import argparse
 
 import sct_utils as sct
-from spinalcordtoolbox.utils import Metavar, SmartFormatter
+from spinalcordtoolbox.utils import Metavar, SmartFormatter, init_sct
 
 # main
 # =======================================================================================================================
@@ -97,6 +97,6 @@ def get_parser():
 # Start program
 #=======================================================================================================================
 if __name__ == "__main__":
-    sct.init_sct()
+    init_sct()
     # call main function
     main()
