+++ scripts/sct_dmri_concat_b0_and_dwi.py
@@ -16,7 +16,7 @@ import argparse
 import numpy as np
 
 from dipy.data.fetcher import read_bvals_bvecs
-from spinalcordtoolbox.utils import Metavar, SmartFormatter
+from spinalcordtoolbox.utils import Metavar, SmartFormatter, init_sct
 from spinalcordtoolbox.image import Image, concat_data
 
 import sct_utils as sct
@@ -153,6 +153,6 @@ def main(args=None):
 
 
 if __name__ == "__main__":
-    sct.init_sct()
+    init_sct()
     # call main function
     main()
