+++ scripts/sct_dmri_concat_bvecs.py
@@ -15,7 +15,7 @@ from __future__ import absolute_import
 import sys, os, argparse
 
 import sct_utils as sct
-from spinalcordtoolbox.utils import Metavar, SmartFormatter
+from spinalcordtoolbox.utils import Metavar, SmartFormatter, init_sct
 
 
 # PARSER
@@ -111,6 +111,6 @@ def main():
 # START PROGRAM
 # ==========================================================================================
 if __name__ == "__main__":
-    sct.init_sct()
+    init_sct()
     # call main function
     main()
