+++ scripts/sct_dmri_transpose_bvecs.py
@@ -27,7 +27,7 @@ from __future__ import absolute_import, division
 import os
 import sys
 import argparse
-from spinalcordtoolbox.utils import Metavar, SmartFormatter
+from spinalcordtoolbox.utils import Metavar, SmartFormatter, init_sct
 
 from sct_utils import extract_fname, printv
 import sct_utils as sct
@@ -87,7 +87,7 @@ def main(args=None):
     fname_in = arguments.bvec
     fname_out = arguments.o
     verbose = int(arguments.v)
-    sct.init_sct(log_level=verbose, update=True)  # Update log level
+    init_sct(log_level=verbose, update=True)  # Update log level
 
     # get bvecs in proper orientation
     from dipy.io import read_bvals_bvecs
@@ -114,6 +114,6 @@ def main(args=None):
 # Start program
 #=======================================================================================================================
 if __name__ == "__main__":
-    sct.init_sct()
+    init_sct()
     # call main function
     main()
