+++ unit_testing/test_qmri.py
@@ -11,8 +11,7 @@ import pytest
 
 from spinalcordtoolbox.qmri import mt
 from spinalcordtoolbox.image import Image
-
-from sct_utils import init_sct
+from spinalcordtoolbox.utils import init_sct
 
 init_sct(log_level=2)  # Set logger in debug mode
 VERBOSE = 0  # Set to 2 to save images, 0 otherwise
